Refactorizando los model de clases remota
