+++ app/src/main/java/com/architect/coders/pokedex/network/PokeService.kt
@@ -1,7 +1,5 @@
 package com.architect.coders.pokedex.network
 
-import com.architect.coders.pokedex.model.PokemonDetail
-import com.architect.coders.pokedex.model.PokemonListResult
 import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.Query
@@ -11,5 +9,5 @@ interface PokeService {
     suspend fun getPokemonList(@Query("offset") offset: Int): PokemonListResult
 
     @GET("pokemon/{name}")
-    suspend fun getPokemonDetail(@Path("name") id: Int): PokemonDetail
+    suspend fun getPokemonDetail(@Path("name") id: Int): PokemonDetailR
 }
