+++ app/src/main/java/com/architect/coders/pokedex/network/Species.kt
@@ -1,9 +1,9 @@
-package com.architect.coders.pokedex.model
+package com.architect.coders.pokedex.network
 
 
 import com.google.gson.annotations.SerializedName
 
-data class TypeX(
+data class Species(
     @SerializedName("name")
     val name: String,
     @SerializedName("url")
