+++ app/src/main/java/com/architect/coders/pokedex/network/StatR.kt
@@ -1,9 +1,9 @@
-package com.architect.coders.pokedex.model
+package com.architect.coders.pokedex.network
 
 
 import com.google.gson.annotations.SerializedName
 
-data class Stat(
+data class StatR(
     @SerializedName("base_stat")
     val baseStat: Int,
     @SerializedName("effort")
